# api/docs/openapi-full.yaml
openapi: 3.0.3
info:
  title: GlobalReach API
  description: Enterprise Email Marketing Platform API
  version: 2.0.0
  contact:
    email: support@globalreach.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.globalreach.com/api/v1
    description: Production server
  - url: https://staging.api.globalreach.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: AUTHENTICATION_FAILED
        message:
          type: string
          example: Invalid credentials
...
    Email:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        to_address:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [pending, sent, delivered, opened, clicked, bounced, converted]
        opened:
          type: boolean
        clicked:
          type: boolean
        converted:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /emails:
    get:
      tags: [Emails]
      summary: Get user emails
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of emails
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Email'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          pages:
                            type: integer
    post:
      tags: [Emails]
      summary: Send email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                campaignId:
                  type: integer
                to:
                  type: string
                subject:
                  type: string
                body:
                  type: string
                templateId:
                  type: integer
                data:
                  type: object
      responses:
        '201':
          description: Email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Email'
  /emails/{id}:
    get:
      tags: [Emails]
      summary: Get email by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Email details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Email'
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /search/emails:
    get:
      tags: [Search]
      summary: Search emails
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Email'
                      total:
                        type: integer
                      page:
                        type: integer
                      pageSize:
                        type: integer
...
  /export/emails/csv:
    get:
      tags: [Export]
      summary: Export emails as CSV
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /export/emails/excel:
    get:
      tags: [Export]
      summary: Export emails as Excel
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
  /export/analytics/pdf:
    get:
      tags: [Export]
      summary: Export analytics as PDF
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
...